# shoes/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample CRUD api for shoes_development db",
        "title": "Sample CRUD api for shoes_development db",
        "contact": {
            "name": "Me",
            "url": "http://me.com/terms.html",
            "email": "me@me.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "xinqi.dev:443",
    "basePath": "/",
    "paths": {
        "/arinternalmetadata_": {
            "get": {
                "description": "GetAllArInternalMetadata_ is a handler to get a slice of record(s) from ar_internal_metadata table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArInternalMetadata_"
                ],
                "summary": "Get list of ArInternalMetadata_",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page requested (defaults to 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of records in a page  (defaults to 20)",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "db sort order column",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PagedResults"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "add to add a single record to ar_internal_metadata table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArInternalMetadata_"
                ],
                "summary": "Add an record to ar_internal_metadata table",
                "parameters": [
                    {
                        "description": "Add ArInternalMetadata_",
                        "name": "ArInternalMetadata_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ArInternalMetadata_"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ArInternalMetadata_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/arinternalmetadata_/{argKey}": {
            "get": {
                "description": "GetArInternalMetadata_ is a function to get a single record from the ar_internal_metadata table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArInternalMetadata_"
                ],
                "summary": "Get record from table ArInternalMetadata_ by  argKey",
                "operationId": "argKey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "argKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ArInternalMetadata_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "ErrNotFound, db record for id not found - returns NotFound HTTP 404 not found error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a single record from ar_internal_metadata table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArInternalMetadata_"
                ],
                "summary": "Update an record in table ar_internal_metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "argKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ArInternalMetadata_ record",
                        "name": "ArInternalMetadata_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ArInternalMetadata_"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ArInternalMetadata_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single record from ar_internal_metadata table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArInternalMetadata_"
                ],
                "summary": "Delete a record from ar_internal_metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "argKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/model.ArInternalMetadata_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "api.PagedResults": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
        "model.ArInternalMetadata_": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "[ 2] created_at                                     datetime             null: false  primary: false  isArray: false  auto: false  col: datetime        len: -1      default: []",
                    "type": "string"
                },
                "key": {
                    "description": "[ 0] key                                            varchar(255)         null: false  primary: true   isArray: false  auto: false  col: varchar         len: 255     default: []",
                    "type": "string"
                },
                "updated_at": {
                    "description": "[ 3] updated_at                                     datetime             null: false  primary: false  isArray: false  auto: false  col: datetime        len: -1      default: []",
                    "type": "string"
                },
                "value": {
                    "description": "[ 1] value                                          varchar(255)         null: true   primary: false  isArray: false  auto: false  col: varchar         len: 255     default: []",
                    "type": "string"
                }
            }
        },
...
    }
}